+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/dialog/o-table-dialog-components.ts
@@ -2,17 +2,13 @@ import { OTableApplyConfigurationDialogComponent } from './apply-configuration/o
 import { OTableExportDialogComponent } from './export/o-table-export-dialog.component';
 import { OTableFilterByColumnDataDialogComponent } from './filter-by-column/o-table-filter-by-column-data-dialog.component';
 import { OTableGroupByColumnsDialogComponent } from './group-by-columns/o-table-group-by-columns-dialog.component';
-import { OTableLoadFilterDialogComponent } from './load-filter/o-table-load-filter-dialog.component';
 import { OTableStoreConfigurationDialogComponent } from './store-configuration/o-table-store-configuration-dialog.component';
-import { OTableStoreFilterDialogComponent } from './store-filter/o-table-store-filter-dialog.component';
 import { OTableVisibleColumnsDialogComponent } from './visible-columns/o-table-visible-columns-dialog.component';
 
 export const O_TABLE_DIALOGS = [
   OTableApplyConfigurationDialogComponent,
   OTableExportDialogComponent,
   OTableStoreConfigurationDialogComponent,
-  OTableLoadFilterDialogComponent,
-  OTableStoreFilterDialogComponent,
   OTableVisibleColumnsDialogComponent,
   OTableFilterByColumnDataDialogComponent,
   OTableGroupByColumnsDialogComponent