+++ projects/ontimize-web-ngx/src/lib/types/o-filter-definition.type.ts
@@ -1,0 +1,4 @@
+export type OFilterDefinition = {
+  name: string;
+  description?: string;
+}